ref(pacscript): extract golang packages into smaller ones
